fix: доработать методы в FilmService.java

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -23,5 +23,5 @@ public class Film {
     int duration;
     @With
     @PositiveOrZero
-    Long likes;
+    Integer likes;
 }
